pyproject.toml
[project]
name = "collabamap"
version = "0.1.0"
description = "its a map that collaborates"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.5.1",
    "beanie[srv]>=1.29.0",
    "fastapi>=0.115.12",
    "fastapi-cors>=0.0.6",
    "itsdangerous>=2.2.0",
    "jose>=1.0.0",
    "openai>=1.69.0",
    "pillow>=11.1.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "starlette>=0.46.1",
    "starsessions>=2.2.1",
    "uvicorn>=0.34.0",
]
